Correct lesson2 and lesson3 and Add lesson4

+++ src/main/java/lesson2/ComplicatedArithmetic.java
@@ -1,14 +1,13 @@
 package lesson2;
 
 public class ComplicatedArithmetic {
-
     public static final double PI = 3.14159265;
 
     public static double circleSquare(int radius){
-        return PI * (double) radius * radius;
+        return PI * radius * radius;
     }
 
     public static double sphereVolume(int radius){
-        return PI * (double) radius * radius * radius * 4 / 3;
+        return PI * radius * radius * radius * 4 / 3;
     }
 }
